app/support/requirements.txt
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
imageio==2.36.0
lazy_loader==0.4
networkx==3.4.2
numpy==1.25.0
opencv-python-headless==4.8.0.74
packaging==24.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==10.0.0
pycparser==2.22
PyMuPDF==1.22.3
pytesseract==0.3.10
PyWavelets==1.7.0
scikit-image==0.20.0
scipy==1.14.1
tifffile==2024.9.20
Wand==0.6.13
